Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultaneous Release. Add a profile that is active by
default so that default behavior is unchanged however allows the
autorelease project to disable building this module.

Issue: RELENG-86
Change-Id: I0999b1794afd1bfcda2baf4054bec212950baf60
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>